backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
langchain==0.0.335
langchain-openai==0.0.2
python-multipart==0.0.6
aiofiles==23.2.1
PyPDF2==3.0.1
httpx==0.25.2

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Security & Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
